Reunify pre- and post-go1.11 identity, and use alternate ways of getting the data
